Move helper methods to a module, adjust time comparison
Trying to be organized
